+++ streamlit_webrtc/models.py
@@ -44,6 +44,11 @@ class VideoProcessorBase(abc.ABC):
         """
         return [self.recv(frames[-1])]
 
+    def on_ended(self):
+        """
+        A callback method which is called when the input track ends.
+        """
+
 
 class VideoTransformerBase(VideoProcessorBase):  # Backward compatiblity
     """
@@ -82,6 +87,11 @@ class AudioProcessorBase(abc.ABC):
                 "`recv_queued` is recommended to use instead."
             )
         return [self.recv(frames[-1])]
+
+    def on_ended(self):
+        """
+        A callback method which is called when the input track ends.
+        """
 
 
 VideoProcessorT = TypeVar("VideoProcessorT", bound=VideoProcessorBase)